requirements.txt
gradio
faiss-cpu
sentence-transformers
openai
numpy<2
torch
transformers